# swagger.yaml
swagger: '2.0'
info:
  description: >-
    WebStarter is a boilerplate project. It's based on Angular Fullstack
    generator. It contains both: client and server.
  version: 1.0.0
  title: WebStarter
  contact:
    email: zangular@yandex.ru
host: '127.0.0.1:9000'
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  AuthSecurity:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /api/users:
    get:
      tags:
        - users
      summary: List all users
      security:
        - AuthSecurity: []
      responses:
        '200':
          description: All users registered
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '401':
          $ref: '#/responses/StdResp401'
    post:
      tags:
        - users
      summary: Add new user
      security:
        - AuthSecurity: []
      parameters:
        - name: new user
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      produces:
        - application/json
        - text/plain
      responses:
        '200':
          description: The user has been added successfully
          schema:
            properties:
              token:
                type: string
        '401':
          $ref: '#/responses/StdResp401'
        '403':
          description: 'something wrong with input data, probably with email'
          schema:
            type: string
            example: The specified email address is already in use.
...
  '/api/users/{userId}':
    parameters:
      - $ref: '#/parameters/userId'
    get:
      tags:
        - users
      summary: Get user by ID
      description: Returns a single user
      security:
        - AuthSecurity: []
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        '401':
          $ref: '#/responses/StdResp401'
        '404':
          description: User not found
    delete:
      tags:
        - users
      summary: Deletes a thing
      security:
        - AuthSecurity: []
      responses:
        '200':
          description: OK
        '401':
          $ref: '#/responses/StdResp401'
        '404':
          description: Not found
...
definitions:
...
  User:
    type: object
    required:
      - name
      - email
      - role
    properties:
      _id:
        type: string
        readOnly: true
      name:
        type: string
      email:
        type: string
      role:
        type: string
responses:
  StdResp401:
    description: UNAUTHORIZED
    schema:
      type: string
parameters:
  userId:
    name: userId
    in: path
    required: true
    type: string